UI class and objects' improvements
A better way to initialize the UI class and an option to receive its pointer at any part of the code. More flags related to scroll boxes and buttons. A new `pane` element which acts as a simple background and a `wall` when clicking GUI objects. Better buttons - flags to only call a function when user releases LMB, option to pass nullptr instead of a function and passing a function that will be called when button is held. README updated.